+++ internal/rules/correctness/ambiguous_returns.go
@@ -59,8 +59,8 @@ func (r *AmbiguousReturnRule) Run(runner *rules.Runner, node ast.Node) {
 	for typ, count := range seen {
 		if count > r.MaxAllowed && !isAllowedDuplicateReturn(typ) {
 			runner.Report(fn.Type.Results.Pos(), rules.Issue{
-				ArgInt1:  count,
-				ArgInt2:  r.MaxAllowed,
+				ArgInt1:  uint32(count),
+				ArgInt2:  uint32(r.MaxAllowed),
 				ID:       rules.AmbiguousReturnID,
 				Severity: r.Severity,
 			})
